fix
NetworkAnimator started throwing this exception:
"Objects are trying to be loaded during a domain backup. This is not allowed as it will lead to undefined behaviour!"
When MPPM was being used. This has to do with the OnBeforeSerialize method and potentially an order of operations issue.
We can accomplish the same serialization within OnValidate, but went ahead and made it virtual in order to provide users with the ability to override and handle their own validation.